chore: add workflow for ray offline tests
