+++ .secrets.baseline
@@ -1333,14 +1333,14 @@
         "filename": "sdk/python/tests/integration/feature_repos/universal/data_sources/file.py",
         "hashed_secret": "d70eab08607a4d05faa2d0d6647206599e9abc65",
         "is_verified": false,
-        "line_number": 258
+        "line_number": 257
       },
       {
         "type": "Secret Keyword",
         "filename": "sdk/python/tests/integration/feature_repos/universal/data_sources/file.py",
         "hashed_secret": "d70eab08607a4d05faa2d0d6647206599e9abc65",
         "is_verified": false,
-        "line_number": 258
+        "line_number": 257
       }
     ],
     "sdk/python/tests/integration/feature_repos/universal/online_store/couchbase.py": [
@@ -1539,5 +1539,5 @@
       }
     ]
   },
-  "generated_at": "2026-03-01T13:20:18Z"
+  "generated_at": "2026-03-01T19:07:31Z"
 }
