+++ pyproject.toml
@@ -282,7 +282,17 @@ python = "~=3.10.0"
 feast = { path = ".", editable = true, extras = ["duckdb", "delta", "grpcio", "test"] }
 
 [tool.pixi.feature.duckdb-tests.tasks]
-test = { cmd = "python -m pytest -n 8 --integration --ignore=sdk/python/tests/integration/offline_store/test_dqm_validation.py sdk/python/tests/integration/offline_store", env = { PYTHONPATH = ".", FULL_REPO_CONFIGS_MODULE = "sdk.python.tests.integration.feature_repos.duckdb_repo_configuration", FEAST_IS_LOCAL_TEST = "True" } }
+test = { cmd = "python -m pytest -n 8 --integration -m 'not ray_offline_stores_only' --ignore=sdk/python/tests/integration/offline_store/test_dqm_validation.py sdk/python/tests/integration/offline_store", env = { PYTHONPATH = ".", FULL_REPO_CONFIGS_MODULE = "sdk.python.tests.integration.feature_repos.duckdb_repo_configuration", FEAST_IS_LOCAL_TEST = "True" } }
+
+[tool.pixi.feature.ray-tests.dependencies]
+python = "~=3.10.0"
+
+[tool.pixi.feature.ray-tests.pypi-dependencies]
+feast = { path = ".", editable = true, extras = ["ray", "grpcio", "test"] }
+
+[tool.pixi.feature.ray-tests.tasks]
+test = { cmd = "python -m pytest -n 8 --integration --ignore=sdk/python/tests/integration/offline_store/test_dqm_validation.py sdk/python/tests/integration/offline_store", env = { PYTHONPATH = ".", FULL_REPO_CONFIGS_MODULE = "sdk.python.tests.integration.feature_repos.ray_repo_configuration", FEAST_IS_LOCAL_TEST = "True" } }
 
 [tool.pixi.environments]
 duckdb-tests = ["duckdb-tests"]
+ray-tests = ["ray-tests"]
