+++ sdk/python/tests/integration/feature_repos/ray_repo_configuration.py
@@ -1,0 +1,6 @@
+from feast.infra.offline_stores.contrib.ray_repo_configuration import (
+    RayDataSourceCreator,
+)
+
+AVAILABLE_OFFLINE_STORES = [("local", RayDataSourceCreator)]
+AVAILABLE_ONLINE_STORES = {"sqlite": ({"type": "sqlite"}, None)}
